adventOfCode/2024/day1/solution_test.go
package day1

import "testing"

func TestShouldReturnProperValuesForPart1TestInput(t *testing.T) {
	testCase := []struct {
		input string
		res   int
	}{

		{
			input: `3   4
4   3
2   5
1   3
3   9
3   3`,
			res: 11,
		},
	}

	for _, tCase := range testCase {
		if res := handlePart1(tCase.input); res != tCase.res {
			t.Fatalf("Expected: %v, got: %v", tCase.res, res)
		}
	}
}

func TestShouldReturnProperValuesForPart1(t *testing.T) {
	res := part1()
	if res != 1151792 {
		t.Fatalf("Got: %v", res)
	}
}

func TestShouldReturnProperValuesForPart2(t *testing.T) {
	res := part2()
	if res != 21790168 {
		t.Fatalf("Got: %v", res)
	}
}


